refactor(gui): split monolithic app.go into domain-specific files
Split the 1427-line app.go into focused domain files:
- app.go (213 lines): Core App struct, interfaces, helper methods
- param.go (286 lines): SSM Parameter Store types and methods
- secret.go (354 lines): Secrets Manager types and methods
- staging.go (617 lines): Staging operation types and methods

This improves maintainability by separating concerns while keeping
all code in the same package to maintain wails bindings compatibility.

Closes #33
